changed regex value from integer to tinyint
changed the value representing the regex flag in the phrase-response system from INTEGER to TINYINT

+++ index.js
@@ -47,7 +47,7 @@ const Phrases = sequelize.define('phrases', {
 	response: Sequelize.TEXT,
 	delete: Sequelize.TINYINT,
 	timeout: Sequelize.INTEGER,
-	regex: Sequelize.INTEGER,
+	regex: Sequelize.TINYINT,
 });
 
 //currently only keeps track of the log channel
